+++ src/Pages/HomePage/HomeComponents/NewArrivalCard.js
@@ -1,13 +1,13 @@
 import React from 'react'
-import '../../Pages/HomePage/HomePage.css'
+import '../HomePage.css'
 
 const NewArrivalCard = ({ imgSrc, title }) => {
     return (
         <div className="img-wrapper pos-rel">
             <div className="grid-txt-overlay pos-abs">
                 <p className="rg-title light-txt center-txt">{title}</p>
-                <a href="./Products/products.html">
-                    <button className="btn subtle-btn mg-t-10">Explore <i className="fa fa-solid fa-arrow-right"></i>
+                <a href="./Products/products.html" >
+                    <button className="btn subtle-btn mg-t-10 center-items">Explore <i className="fa fa-solid fa-arrow-right"></i>
                     </button>
                 </a>
             </div>
