+++ src/Pages/HomePage/HomePage.css
@@ -81,3 +81,23 @@
   color: var(--orange-color);
   border: 1px solid transparent;
 }
+
+.category-preview {
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  width: min-content;
+}
+.round-img-container {
+  width: 15rem;
+  height: 15rem;
+  border: 2px solid transparent;
+  border-radius: 50%;
+}
+
+.round-img-container:hover {
+  cursor: pointer;
+  border: 2px solid var(--violet-color);
+}
