+++ src/Pages/HomePage/HomePage.js
@@ -1,14 +1,8 @@
 import React from 'react'
 import "./HomePage.css"
-
 import Navbar from "../../Components/Navbar/Navbar"
 import Footer from "../../Components/Footer/Footer"
-
-import DealsSection from './PageSections/DealsSection';
-import BrandSection from './PageSections/BrandsSection.js'
-import ArrivalsSection from './PageSections/ArrivalsSection.js'
-import CategoriesSection from './PageSections/CategoriesSection';
-import HeroSection from './PageSections/HeroSection'
+import { DealsSection, BrandSection, ArrivalsSection, CategoriesSection, HeroSection } from './PageSections/index'
 
 const HomePage = () => {
     return (
