+++ src/Pages/HomePage/PageSections/index.js
@@ -1,0 +1,23 @@
+import DealsSection from './DealsSection';
+import BrandSection from './BrandsSection.js'
+import ArrivalsSection from './ArrivalsSection.js'
+import CategoriesSection from './CategoriesSection';
+import HeroSection from './HeroSection'
+
+import CategoryCard from '../HomeComponents/CategoryCard'
+import CyberPunk from "../../../Assets/Categories/cyberpunk.jfif"
+import Headphone from "../../../Assets/Categories/headphone.jpg"
+import Hoodie from "../../../Assets/Categories/hoodie.png"
+import PS5 from "../../../Assets/Categories/ps5.jpg"
+
+import XiaoHome from "../../../Assets/HomeImages/xiao-genshin.jfif"
+import CollectiblesVal from "../../../Assets/HomeImages/collectibles-val.jfif"
+
+
+import Tencent from "../../../Assets/Brands/tencent.png"
+import Sony from "../../../Assets/Brands/sony.png"
+import Nvidea from "../../../Assets/Brands/nvidea.png"
+import Intel from "../../../Assets/Brands/intel.png"
+import EA from "../../../Assets/Brands/electronics-arts.png"
+
+export { DealsSection, BrandSection, ArrivalsSection, CategoriesSection, HeroSection, CategoryCard, CyberPunk, Headphone, Hoodie, PS5, XiaoHome, CollectiblesVal, Tencent, Sony, EA, Intel, Nvidea }
